Removed a bunch of warnings thrown by gcc, when compiled with CMAKE_BUILD_TYPE=Debug.
Note that there are still a lot of warnings left. which become apparent when switching
to CMAKE_BUILD_TYPE=Release. Namely those are 'may be used uninitialized' warnings.

+++ src/Geometry/Mesh.C
@@ -306,13 +306,13 @@ void Mesh::hashTriFaces()
   }
   std::vector<int> BucketCount(3*n_points);
   
-  for(int i=0;i<triangle.size();++i)
+  for(size_t i=0;i<triangle.size();++i)
   {
     int hash = triangle[i].nodes[0]+triangle[i].nodes[1]+triangle[i].nodes[2];
     (BucketCount[hash])++;
   }
 
-  for(int i=0;i<triangle.size();++i)
+  for(size_t i=0;i<triangle.size();++i)
   {
     int hash = triangle[i].nodes[0]+triangle[i].nodes[1]+triangle[i].nodes[2];
 
@@ -491,7 +491,7 @@ void Mesh::createInnerFaces()
   for(unsigned int i=0; i<localHash.size(); i++)
     localHash[i].resize(0);
   
-  for(int i=0;i<triangle.size();++i)
+  for(size_t i=0;i<triangle.size();++i)
   {
     int hash = triangle[i].nodes[0]+triangle[i].nodes[1]+triangle[i].nodes[2];
